+++ HISTORY.md
@@ -1,3 +1,7 @@
+
+==================
+* Efficiency upgrade for infinite version
+
 v1.2.0
 ==================
 * Polymer 1.x/2.x hybrid element support